ir-server/app/requirements.txt
paho-mqtt
homeassistant-mqtt-binding

